Jesus save me from linking these hell-spawned classes together
So; Long story short, everything broke on me and I spent literally all
night fixing magical floating errors trying to connect our layers
together.

Let's all agree that whenever we change something in a class we test it
before linking it together.

So; UI got it's own class now and can communicate to the logic layer,
the logic layer works by black magic and can communicate to the Data
layer. Everyone seems to be playing nicely together and the software
seems to work A-ok for the most part.

If we want to add, fix or change anything we really should be careful.
This thing has proven itself to be ~very~ sensitive to minor changes.

I'm going to sleep. wake me up if the world is ending.