FileData support for Computers started
Ok, a bit update:

*I've created the "Computer" class to act as a mirror to
ComputerScientst. It will eventually work in the exact same way but...
as a computer.

* The backend is a mess currently, there is a lot to do before this
becomes remotly functional again

* tried to add a few documentations to a couple of cpp files

* FileData::Add works. So, if you want to add something to the SQL
database using c++ then FileData will be able to add both
computerscientists and computers

* I'll work more on this later this evening, but just be aware that the
program does not work. Anything you add or change will have to be tested
on its own.

* If you look over what changes I made you might see past my spaghetty
madness and see what I was intending, but if not please ask. We can't
work on the code if we can't understand it so if anything confuses you
then ask without hesitation.