Search Connections seems to be ok
I hope, it's hard to tell. It's dark, I'm tired, and I see nothing but a
blend of code, Reddit and the waterbottle next to me. My dark magic has
been long and slow, maybe in part how lazy I've been. I opened the
window in hopes to aid my magical conjuring but all I got was cold. Dark
magic is a tricky buisness, you never know quite what you get. For
instance, while I can search connections fine from my end maybe the dark
powers are limited to this machine only, leaving a red sea of errors for
all others. We'll see come morning, when the darkness of the night
yields for the light and day shines upone us. By then dark magic will be
much harder to perform; but I'm not a master of the code for no reason.
I'll perform my deed in light and day, and hopefully I'll be more awake
and actually productive in the process.

Whatever the result is, it will have to wait to morning. Morning, and
not a moment sooner.